test(doctest): consolidate Markdown coverage in plugin suite
