+++ packages/tools/doctest/test/fixtures/markdown/docs/typed.md
@@ -1,9 +1,9 @@
 # Typed Markdown examples
 
 This prose is not JavaScript and must never execute.
-throw new Error("R11_MARKDOWN_PROSE_EXECUTED")
+throw new Error("Markdown prose must not execute")
 
-```ts import.meta.vitest name=static-first
+```ts import.meta.vitest name=static-import
 import { expect } from "vitest"
 import { answer } from "./helper.ts"
 const local: number = answer
@@ -12,11 +12,18 @@ expect(local).toBe(42)
 expect(import.meta.url).toContain("/docs/typed.md")
 ```
 
-```ts import.meta.vitest name=dynamic-second
+```ts import.meta.vitest name=dynamic-import
 import { expect } from "vitest"
 const { answer } = await import("./helper.ts")
 const local: number = answer + 1
 expect.assertions(2)
 expect(local).toBe(43)
 expect(import.meta.url).toContain("/docs/typed.md")
+```
+
+```ts import.meta.vitest name=javascript
+import { expect } from "vitest"
+const local = 42
+expect.assertions(1)
+expect(local).toBe(42)
 ```
