--- packages/tools/doctest/test/fixtures/markdown/helper.ts
@@ -1,2 +1,0 @@
-export const answer = -1
-throw new Error("R11_ROOT_DECOY_EXECUTED")
